+++ src/main/java/com/hedvig/notificationService/customerio/web/EventController.kt
@@ -1,6 +1,6 @@
 package com.hedvig.notificationService.customerio.web
 
-import com.hedvig.notificationService.customerio.EventHandler
+import com.hedvig.notificationService.service.event.EventHandler
 import com.hedvig.notificationService.customerio.dto.ChargeFailedEvent
 import com.hedvig.notificationService.customerio.dto.ContractCreatedEvent
 import com.hedvig.notificationService.customerio.dto.ContractRenewalQueuedEvent
