+++ src/main/java/com/hedvig/notificationService/service/request/EventRequestHandler.kt
@@ -1,6 +1,6 @@
 package com.hedvig.notificationService.service.request
 
-import com.hedvig.notificationService.customerio.EventHandler
+import com.hedvig.notificationService.service.event.EventHandler
 import com.hedvig.notificationService.customerio.dto.ChargeFailedEvent
 import com.hedvig.notificationService.customerio.dto.ContractCreatedEvent
 import com.hedvig.notificationService.customerio.dto.ContractRenewalQueuedEvent
