+++ src/test/java/com/hedvig/notificationService/customerio/web/ContractCreatedEventCreatesContractTest.kt
@@ -7,7 +7,7 @@ import assertk.assertions.extracting
 import assertk.assertions.isNotNull
 import com.hedvig.notificationService.customerio.ConfigurationProperties
 import com.hedvig.notificationService.customerio.CustomerioService
-import com.hedvig.notificationService.customerio.EventHandler
+import com.hedvig.notificationService.service.event.EventHandler
 import com.hedvig.notificationService.customerio.dto.ContractCreatedEvent
 import com.hedvig.notificationService.customerio.hedvigfacades.MemberServiceImpl
 import com.hedvig.notificationService.customerio.state.CustomerioState
