+++ src/test/java/com/hedvig/notificationService/customerio/web/OnQuoteCreatedEventTest.kt
@@ -1,7 +1,7 @@
 package com.hedvig.notificationService.customerio.web
 
 import com.hedvig.notificationService.customerio.CustomerioService
-import com.hedvig.notificationService.customerio.EventHandler
+import com.hedvig.notificationService.service.event.EventHandler
 import com.hedvig.notificationService.customerio.builders.EMAIL
 import com.hedvig.notificationService.customerio.builders.MEMBER_ID
 import com.hedvig.notificationService.customerio.builders.SSN
